auth_service/auth_service/auth_app/requirements.txt
django==3.2.5
djangorestframework==3.12.4
mongoengine==0.24.1
pymongo==3.12.0
rabbitmq==3.8.3
channels==3.0.4
channels_redis==3.3.1
djangorestframework-jwt==1.11.0
swagger==2.2.3